+++ homework1/jsconfig.json
@@ -1,0 +1,7 @@
+{
+    "compilerOptions": {
+      "baseUrl": "./src",
+      "checkJs": true,
+      "jsx": "react"
+    }
+  }
